live: move CPIO functions into io
