+++ src/io/mod.rs
@@ -19,12 +19,14 @@ use std::io::{ErrorKind, Read, Write};
 mod bls;
 mod compress;
 mod hash;
+mod initrd;
 mod limit;
 mod verify;
 
 pub use self::bls::*;
 pub use self::compress::*;
 pub use self::hash::*;
+pub use self::initrd::*;
 pub use self::limit::*;
 pub use self::verify::*;
